Add files via upload
I had to do a lot of adjusting to make my code work. I tried to compare the data structure of the geojson in my project to the example and found it did not produce the same results. In order to make things work overall, I had to pull back some functions that were nested together. This helped functions interact with each other better, which helped get to the end result. In regards to how I wanted to order the buttons, I wanted to use the years each accident occurred rather than the number of fatalities shown (since that does show data in a time-space form). Due to the years being added, I had to include a lot of different structure to properly implement the years into the code. This brought challenges, since I needed to use an array for a singular value and the code provided in the example had a different data structure. I had referenced stackoverflow mainly for information on how to utilize this. However I was a little stuck. Therefore I went to ChatGPT to offer some assistance. I am always hesitant to use AI in any sort of educational context, but I have read the policy and used it more as a tool to help me push over my coding block. Overall, this project took a lot of time out of my week, but I believe it will put me in a better position for the future! I really did have to shuffle a lot from my initial structure, so all of that took up a lot of time. For future projects (and due to me not having enough time) I will adjust the code to include the arrow pngs and make the map look more official for the final project. I am an honest student and again, I never feel right using AI for any educational attainment. Regardless, coding can get me stuck at times. I am thankful for your policy, as I will use it as a tool to better my understanding as a student!